// src/payload-types.ts
export interface Document {
  id: string;
  displayName?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
// ...
export interface BillingPlan {
  id: string;
  name: string;
  planId: string;
  price: number;
  price_formatted: string;
  interviewCount?: number | null;
  interviewDuration: number;
  freeTrialPeriod?: number | null;
  description?: string | null;
  features?:
    | {
        feature?: string | null;
        id?: string | null;
      }[]
    | null;
  buy_now_url: string;
  planDuration?: number | null;
  expiredIn?: number | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "users".
 */
export interface User {
  id: string;
  roles?: ("platform-admin" | "platform-user")[] | null;
  name: string;
  phone?: string | null;
  language?: string | null;
  inviteToken: string;
  invitedUsers?: (string | User)[] | null;
  group?: (string | null) | Group;
  registeredAt?: string | null;
  reportAccess?: boolean | null;
  onboarding?: {
    persona?: ("student" | "professional" | "free") | null;
    industry?: (string | null) | Industry;
    targetRole?: {
      level?: (string | null) | Level;
      role?: (string | null) | Role;
    };
    challenge?: (string | Challenge)[] | null;
    trainingPlan?: {
      trainingType?: ("structured" | "freestyle") | null;
      hours?: number | null;
      mins?: number | null;
      per?: ("day" | "week" | "month") | null;
    };
    interviewDate?: string | null;
    cv?: (string | null) | Document;
  };
  freeTrialStartsAt?: string | null;
  freeTrialExpiresAt?: string | null;
  interviewPlan?: (string | null) | BillingPlan;
  interviewSubscriptionId?: string | null;
  interviewSubscriptionStatus?: string | null;
  interviewSubscriptionStartsAt?: string | null;
  interviewSubscriptionExpiresAt?: string | null;
  consultationPlan?: (string | null) | BillingPlan;
  consultationSubscriptionId?: string | null;
  consultationSubscriptionStatus?: string | null;
  consultationSubscriptionStartsAt?: string | null;
  consultationSubscriptionExpiresAt?: string | null;
  planUsage?: {
    usedInterviewTime?: number | null;
    usedConsultation?: number | null;
    usedConsultationTime?: number | null;
  };
  social?: {
    issuer?: ("google" | "facebook" | "linkedin" | "") | null;
    userId?: string | null;
  };
  picture?: string | null;
  entity?: {
    tenant?: (string | null) | Tenant;
    role?: ("owner" | "staff") | null;
  };
  notifications?:
    | {
        [k: string]: unknown;
      }
    | unknown[]
    | string
    | number
    | boolean
    | null;
  updatedAt: string;
  createdAt: string;
  enableAPIKey?: boolean | null;
  apiKey?: string | null;
  apiKeyIndex?: string | null;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  _verified?: boolean | null;
  _verificationToken?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  password?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "industries".
 */
export interface Industry {
  id: string;
  label: string;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "levels".
 */
export interface Level {
  id: string;
  label: string;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "roles".
 */
export interface Role {
  id: string;
  label: string;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "challenges".
 */
export interface Challenge {
  id: string;
  label: string;
  value: string;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "goals".
 */
export interface Goal {
  id: string;
  label: string;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "tenants".
 */
export interface Tenant {
  id: string;
  name: string;
  type?: (string | null) | InstitutionType;
  slug?: string | null;
  fromName?: string | null;
  fromAddress?: string | null;
  transportOptions?: {
    host?: string | null;
    port?: number | null;
    user?: string | null;
    pass?: string | null;
    secure?: boolean | null;
  };
  template?: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ("ltr" | "rtl") | null;
      format: "left" | "start" | "center" | "right" | "end" | "justify" | "";
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  status?: boolean | null;
  public?: boolean | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "institutionTypes".
 */
export interface InstitutionType {
  id: string;
  label: string;
  updatedAt: string;
  createdAt: string;
}
// ...
export interface Group {
  id: string;
  name: string;
  creator: string | User;
  updatedAt: string;
  createdAt: string;
}